Add test to ensure iidxes kept in sync
